Move AnimationMode to own package
